app/requirements.txt
# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
psycopg2-binary>=2.9.0  # PostgreSQL support
gunicorn>=21.0.0  # Production server
dj-database-url>=2.0.0  # Database URL parsing

# System monitoring
psutil>=5.9.0

# Existing ML dependencies (from parent requirements.txt)
numpy>=1.21.0,<2.0.0
pandas>=1.3.0,<2.0.0
scikit-learn>=1.0.0,<2.0.0
xgboost>=1.5.0,<2.0.0
lightgbm>=3.3.0,<4.0.0
imbalanced-learn>=0.9.0,<1.0.0

# Deep Learning
torch>=2.0.0,<3.0.0
transformers>=4.20.0,<5.0.0

# Template Mining
drain3>=0.9.0

# Utilities
python-dotenv>=1.0.0
whitenoise>=6.5.0  # Static file serving

# Optional: Celery for async tasks
# celery>=5.3.0
# redis>=4.5.0
